+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 1.0a13 (unreleased)
 -------------------
 
-- Nothing changed yet.
+Bugfixes:
+
+- Don't fetch unnecessary PasswordResetTool in Plone 5.1
+  [tomgross]
 
 
 1.0a12 (2017-04-03)