+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 
 Bugfixes:
 
+- Don't fetch unnecessary PasswordResetTool in Plone 5.1
+  [tomgross]
+
 - Add the title to the workflow history in the @workflow endpoint. This fixes
   #279.
   [sneridagh]